# docs/swagger.yaml
definitions:
...
  todolist_internal_dto.AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  todolist_internal_dto.AuthResponse:
    properties:
      expires_at:
        type: string
      refresh_expires_at:
        type: string
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/todolist_internal_dto.UserResponse'
    type: object
  todolist_internal_dto.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  todolist_internal_dto.CreatePersonRequest:
    properties:
      birth_date:
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 20
        minLength: 11
        type: string
      tax_id:
        maxLength: 14
        minLength: 11
        type: string
    required:
    - email
    - name
    - tax_id
    type: object
...
  todolist_internal_dto.CreateUserRequest:
    properties:
      password:
        minLength: 8
        type: string
      person:
        $ref: '#/definitions/todolist_internal_dto.CreatePersonRequest'
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - person
    - username
    type: object
  todolist_internal_dto.ErrorInfo:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        example: Invalid input data
        type: string
    type: object
...
  todolist_internal_dto.PersonInfo:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  todolist_internal_dto.PersonResponse:
    properties:
      birth_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      tax_id:
        type: string
      updated_at:
        type: string
    type: object
  todolist_internal_dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/todolist_internal_dto.ErrorInfo'
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  todolist_internal_dto.UpdatePersonRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 20
        minLength: 11
        type: string
    type: object
...
  todolist_internal_dto.UserResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      person:
        $ref: '#/definitions/todolist_internal_dto.PersonInfo'
      person_id:
        type: integer
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Password change data
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/todolist_internal_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get access token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/todolist_internal_dto.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      summary: Login user
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/todolist_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      summary: Register a new user
      tags:
      - auth
  /api/v1/people:
    post:
      consumes:
      - application/json
      description: Create a new person record
      parameters:
      - description: Person data
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/todolist_internal_dto.CreatePersonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.PersonResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      summary: Create a new person
      tags:
      - people
  /api/v1/people/{id}:
    get:
      consumes:
      - application/json
      description: Get person details by ID
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.PersonResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Get person by ID
      tags:
      - people
    put:
      consumes:
      - application/json
      description: Update person details
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: string
      - description: Person data
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/todolist_internal_dto.UpdatePersonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.PersonResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Update person
      tags:
      - people
...
swagger: "2.0"